requirements.txt
tavily-python>=0.3.0
mcp[cli]>=1.0.0
mcp>=0.0.5
dependency-injector
debugpy>=1.6.0
gunicorn>=21.2.0
fastapi>=0.100.0
uvicorn>=0.23.1
langchain>=0.1.0
langchain-community>=0.0.10  # Add this line
langchain-core>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
faiss-cpu>=1.8.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.1.1
requests
tavily-python
sentence-transformers>=2.5.0
openai>=1.0.0
tqdm>=4.65.0
torch>=2.0.0
torchvision>=0.15.0
